Unrelated: move DecisionsAndContext
Avoid a circular import (that we didn't see because luck)
